requirements.txt
git+https://github.com/Rapptz/discord.py@master
jishaku
pillow-SIMD